+++ tests/Corvus.Text.Json.Tests.GeneratedModels.NullOrUndefinedExceptNonNullDefaulted/ObjectWithDefaultArray.cs
@@ -1,0 +1,6 @@
+namespace Corvus.Text.Json.Tests.GeneratedModels.NullOrUndefinedExceptNonNullDefaulted.Draft202012;
+
+[JsonSchemaTypeGenerator("./object-with-default-array-2020-12.json")]
+public readonly partial struct ObjectWithDefaultArray
+{
+}